Changed question 3 to output log files in console

+++ question-3/question3.js
@@ -12,6 +12,7 @@ const makeLogs = () => {
     process.chdir(target);
     for (let i = 0; i < 10; i++) {
         fs.writeFileSync(`log${i}.txt`, `This is log file ${i}`);
+        console.log(`log${i}.txt`, `This is log file ${i}`)
     }
 }
 
@@ -19,7 +20,13 @@ const deleteLogs = () => {
     const target = './logs';
     if (fs.existsSync(target)) {
         fs.readdirSync(target).forEach((file) => {
-            fs.unlinkSync(`${target}/${file}`);
+            for (let i = 0; i < 10; i++) {
+                if (file === `log${i}.txt`) {
+                    fs.unlinkSync(`${target}/${file}`);
+                    console.log(`delete files ${file}`)
+                }
+            }
+            //fs.unlinkSync(`${target}/${file}`);
         });
         fs.rmdirSync(target);
     }
